+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/controllers/EditorController.java
@@ -25,6 +25,7 @@ import com.onlyoffice.integration.documentserver.models.enums.Action;
 import com.onlyoffice.integration.documentserver.storage.FileStoragePathBuilder;
 import com.onlyoffice.integration.entities.User;
 import com.onlyoffice.integration.dto.Mentions;
+import com.onlyoffice.integration.dto.UserInfo;
 import com.onlyoffice.integration.documentserver.models.enums.Type;
 import com.onlyoffice.integration.documentserver.models.filemodel.FileModel;
 import com.onlyoffice.integration.services.UserServices;
@@ -117,6 +118,8 @@ public class EditorController {
         }
 
         User user = optionalUser.get();
+        user.setImage(user.getAvatar() ? storagePathBuilder.getServerUrl(true) + "/css/img/uid-"
+                + user.getId() + ".png" : null);
 
         // get file model with the default file parameters
         FileModel fileModel = fileConfigurer.getFileModel(
@@ -150,6 +153,8 @@ public class EditorController {
 
         // get user data for mentions and add it to the model
         model.addAttribute("usersForMentions", getUserMentions(uid));
+
+        model.addAttribute("usersInfo", getUsersInfo(uid));
         return "editor.html";
     }
 
@@ -167,6 +172,19 @@ public class EditorController {
         }
 
         return usersForMentions;
+    }
+
+    private List<UserInfo> getUsersInfo(final String uid) {  // get user data for mentions
+        List<UserInfo> usersInfo = new ArrayList<>();
+        if (uid != null && !uid.equals("4")) {
+            List<User> list = userService.findAll();
+            for (User u : list) {
+                String image = u.getAvatar() ? storagePathBuilder.getServerUrl(true) + "/css/img/uid-"
+                + u.getId() + ".png" : null;
+                usersInfo.add(new UserInfo(u.getId(), u.getName(), u.getEmail(), image));
+            }
+        }
+        return usersInfo;
     }
 
     @SneakyThrows